refactor(logic): co-locate overture stage next/previous (Refs #2391)
Move nextOvertureStage and previousStage into diplomacy/overture_stage_navigation.dart
so suggestion and resolution share one chain definition (issue Pattern 2).

Re-export from order_suggestion_context for existing imports. Add tests that
previousStage inverts nextOvertureStage along the progression chain.
