+++ packages/colonizethis_logic/lib/src/diplomacy/overture_resolver.dart
@@ -5,6 +5,7 @@ import '../constants.dart';
 import '../dossier/evidence_rules.dart';
 import '../turn/turn_resolution_result.dart';
 import 'diplomacy_resolver.dart';
+import 'overture_stage_navigation.dart';
 
 Game appendDiplomaticEvent(
   Game game,
@@ -377,21 +378,6 @@ OverturePaymentsResult processOverturePayments(
 
   state = state.copyWith(players: players, overtureStates: overtures);
   return OverturePaymentsResult(state);
-}
-
-OvertureStage previousStage(OvertureStage stage) {
-  switch (stage) {
-    case OvertureStage.tradeConsulate:
-      return OvertureStage.none;
-    case OvertureStage.embassy:
-      return OvertureStage.tradeConsulate;
-    case OvertureStage.nap:
-      return OvertureStage.embassy;
-    case OvertureStage.joinEmpire:
-      return OvertureStage.nap;
-    case OvertureStage.none:
-      return OvertureStage.none;
-  }
 }
 
 Game advanceOvertures(Game game, int turn) {
